# openapi.json
{
  "components": {
    "schemas": {
...
      "AttributeFindingRead": {
        "description": "Attribution detail \u2014 AttributeRead enriched with candidates + per-item evidence.",
        "properties": {
          "art9": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Art9"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "evidence_items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRead"
            },
            "title": "Evidence Items",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "neighborhood": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Neighborhood"
          },
          "precision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Precision"
          },
          "reasoning": {
            "title": "Reasoning",
            "type": "string"
          },
          "reliability": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Reliability"
              },
              {
                "type": "null"
              }
            ]
          },
          "sensitive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensitive"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "text_only_reliability": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Reliability"
              },
              {
                "type": "null"
              }
            ]
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "code",
          "label",
          "value",
          "reliability",
          "severity",
          "reasoning",
          "candidates",
          "evidence_items"
        ],
        "title": "AttributeFindingRead",
        "type": "object"
      },
      "AttributeRead": {
        "description": "A dashboard attribute card. `null` reliability \u21d4 abstain.",
        "properties": {
          "abstain": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abstain"
          },
          "art9": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Art9"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "evidence": {
            "title": "Evidence",
            "type": "string"
          },
          "evidence_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Count"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "reliability": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Reliability"
              },
              {
                "type": "null"
              }
            ]
          },
          "sensitive": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sensitive"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "code",
          "label",
          "value",
          "detail",
          "reliability",
          "evidence",
          "severity"
        ],
        "title": "AttributeRead",
        "type": "object"
      },
      "BBox": {
        "properties": {
          "h": {
            "title": "H",
            "type": "number"
          },
          "w": {
            "title": "W",
            "type": "number"
          },
          "x": {
            "title": "X",
            "type": "number"
          },
          "y": {
            "title": "Y",
            "type": "number"
          }
        },
        "required": [
          "x",
          "y",
          "w",
          "h"
        ],
        "title": "BBox",
        "type": "object"
      },
...
      "Candidate": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          }
        },
        "required": [
          "rank",
          "label",
          "note"
        ],
        "title": "Candidate",
        "type": "object"
      },
...
      "EvidenceRead": {
        "description": "One piece of evidence \u2014 `proven` (causal ablation) or `likely` (correlational).",
        "properties": {
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "citation": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citation"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "exif": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Exif"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "proven",
              "likely"
            ],
            "title": "Kind",
            "type": "string"
          },
          "marginal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marginal"
          },
          "proxy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proxy"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "spans": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spans"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "type": {
            "enum": [
              "text",
              "photo"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "type",
          "source",
          "date",
          "rationale"
        ],
        "title": "EvidenceRead",
        "type": "object"
      },
      "Exif": {
        "properties": {
          "device": {
            "title": "Device",
            "type": "string"
          },
          "gps": {
            "title": "Gps",
            "type": "string"
          },
          "place": {
            "title": "Place",
            "type": "string"
          },
          "taken": {
            "title": "Taken",
            "type": "string"
          }
        },
        "required": [
          "gps",
          "place",
          "device",
          "taken"
        ],
        "title": "Exif",
        "type": "object"
      },
...
      "InferenceConfirm": {
        "description": "POST /v1/inferences/{id}/confirm \u2014 the user confirms/denies (a live ground-truth signal).",
        "properties": {
          "confirmed": {
            "title": "Confirmed",
            "type": "boolean"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value",
          "confirmed"
        ],
        "title": "InferenceConfirm",
        "type": "object"
      },
      "Problem": {
        "description": "RFC 9457 problem+json error body (the one error shape across the API).",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
      "Reliability": {
        "description": "Calibrated reliability \u2014 point estimate + interval, 0..1 (the UI renders %).",
        "properties": {
          "hi": {
            "title": "Hi",
            "type": "number"
          },
          "lo": {
            "title": "Lo",
            "type": "number"
          },
          "point": {
            "title": "Point",
            "type": "number"
          }
        },
        "required": [
          "point",
          "lo",
          "hi"
        ],
        "title": "Reliability",
        "type": "object"
      },
      "RemediationCreate": {
        "description": "POST /v1/inferences/{id}/remediations \u2014 decoy requires per-use consent.",
        "properties": {
          "decoy": {
            "default": false,
            "title": "Decoy",
            "type": "boolean"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy"
          }
        },
        "title": "RemediationCreate",
        "type": "object"
      },
...
      "RunAccepted": {
        "description": "202 response \u2014 the client polls GET /v1/runs/{id} until terminal.",
        "properties": {
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunAccepted",
        "type": "object"
      },
...
      "Severity": {
        "description": "Per-persona severity matrix; the UI computes `balanced = max(atrisk, jobseeker)`.",
        "properties": {
          "atrisk": {
            "enum": [
              "low",
              "moderate",
              "high",
              "extreme"
            ],
            "title": "Atrisk",
            "type": "string"
          },
          "jobseeker": {
            "enum": [
              "low",
              "moderate",
              "high",
              "extreme"
            ],
            "title": "Jobseeker",
            "type": "string"
          }
        },
        "required": [
          "atrisk",
          "jobseeker"
        ],
        "title": "Severity",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Glasshouse",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/inferences": {
      "get": {
        "operationId": "list_inferences_v1_inferences_get",
        "parameters": [
          {
            "in": "query",
            "name": "run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id"
            }
          },
          {
            "in": "query",
            "name": "profile_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile Id"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AttributeRead"
                  },
                  "title": "Response List Inferences V1 Inferences Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Implemented"
          }
        },
        "summary": "List Inferences",
        "tags": [
          "inferences"
        ]
      }
    },
    "/v1/inferences/{inference_id}": {
      "get": {
        "description": "Ranked candidates with calibrated reliability + the evidence join.",
        "operationId": "get_inference_v1_inferences__inference_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "inference_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inference Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributeFindingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Get Inference",
        "tags": [
          "inferences"
        ]
      }
    },
    "/v1/inferences/{inference_id}/confirm": {
      "post": {
        "operationId": "confirm_inference_v1_inferences__inference_id__confirm_post",
        "parameters": [
          {
            "in": "path",
            "name": "inference_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inference Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InferenceConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttributeFindingRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Confirm Inference",
        "tags": [
          "inferences"
        ]
      }
    },
    "/v1/inferences/{inference_id}/remediations": {
      "post": {
        "description": "Trigger a remediation run; decoy requires per-use consent. Lands at M3.",
        "operationId": "create_remediation_v1_inferences__inference_id__remediations_post",
        "parameters": [
          {
            "in": "path",
            "name": "inference_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Inference Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemediationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Unprocessable Entity"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {}
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Create Remediation",
        "tags": [
          "inferences"
        ]
      }
    },
...
  }
}